# fsync all the things