--- a/include/git2/common.h
+++ b/include/git2/common.h
@@ -179,6 +179,7 @@ typedef enum {
 	GIT_OPT_SET_SSL_CIPHERS,
 	GIT_OPT_GET_USER_AGENT,
 	GIT_OPT_ENABLE_OFS_DELTA,
+	GIT_OPT_ENABLE_SYNCHRONIZED_OBJECT_CREATION,
 } git_libgit2_opt_t;
 
 /**
@@ -315,6 +316,13 @@ typedef enum {
  *		> and thus smaller resultant packfiles.
  *		> Packfiles containing offset deltas can still be read.
  *		> This defaults to enabled.
+ *
+ *	* opts(GIT_OPT_ENABLE_SYNCHRONIZED_OBJECT_CREATION, int enabled)
+ *
+ *		> Enable synchronized writes of new objects using `fsync`
+ *		> (or the platform equivalent) to ensure that new object data
+ *		> is written to permanent storage, not simply cached.  This
+ *		> defaults to disabled.
  *
  * @param option Option key
  * @param ... value to set the option