--- a/src/object.c
+++ b/src/object.c
@@ -16,6 +16,7 @@
 #include "tag.h"
 
 bool git_object__strict_input_validation = true;
+bool git_object__synchronized_writing = false;
 
 typedef struct {
 	const char	*str;	/* type name string */