--- a/src/object.h
+++ b/src/object.h
@@ -10,6 +10,7 @@
 #include "repository.h"
 
 extern bool git_object__strict_input_validation;
+extern bool git_object__synchronized_writing;
 
 /** Base git object for inheritance */
 struct git_object {