… src/repository.h
… src/settings.c
… src/unix/posix.h
… src/win32/posix_w32.c
… tests/odb/loose.c
… tests/pack/packbuilder.c
… tests/refs/create.c