Optionally `fsync` loose objects, packfiles and their indexes, loose references and packed reference files.  This can be enabled by setting the option `GIT_OPT_ENABLE_SYNCHRONIZED_OBJECT_CREATION`, and it is not enabled by default.

Additionally, enable the `do_fsync` option to the loose odb backend to perform similar calls.

At present, we do not honor `core.fsyncObjectFiles`, though this would be a logical next step.