@ethomson 2016-12-27
Fixed this up for windows; should be ready for review.