@ioquatix 2017-02-13
Are you calling fsync on the parent directory when creating new files? Is this necessary?