@ethomson 2017-02-13
> Are you calling fsync on the parent directory when creating new files?

No - good call.  On at least POSIX-type systems this is necessary when creating new files.  As far as I can tell, you cannot `FlushFileBuffers` on a directory handle on win32.