@ioquatix 2017-02-13
Once you commit this into master and expose it via rugged (Ruby) somehow, I will test it and let you know how it affects performance.