@ioquatix 2017-02-13
> No - good call. On at least POSIX-type systems this is necessary when creating new files.

One more case - after calling `link` to "move" data into place, should we be calling fsync on the parent directory to ensure the results of `link` are persisted to disk?