@ioquatix 2017-02-13
I'd like to see this PR get merged, but I'd also like to start an extended discussion on how to make this more efficient.

Of course, calling `fsync` after every write, is not efficient. It will introduce significant latency. For many use cases, `fsync` only needs to be called before updating the target ref.

I think it would be nice to review how Apache Lucene implements durability: http://blog.mikemccandless.com/2014/04/testing-lucenes-index-durability-after.html

In particular, the way they batch changes, and test that fsync works seem like good ideas suitable for the odb backend.

I'm not suggesting we defer this PR at all - correct is more important than performance. But, we should probably also consider performance too. If we could make the `fsync` code path almost as fast as the alternative, and it avoids corruption, I'd humbly propose it becomes the default.