@pureconfig 2017-02-14
@ioquatix 

> I'd like to see this PR get merged, but I'd also like to start an extended discussion on how to make this more efficient.

This is why git doesn't call fsync, it kills performance.  Databases solves this by having a single file for the undo/redo logs, which is truly append only (which the SATA NCQ likes and optimizes.)

I guess this is why git is not a concurrent database system, and never will be.