@pureconfig 2017-02-15
@ethomson Don't use FlushFileBuffers, it can be extremely slow. Use FILE_FLAG_WRITE_THROUGH | FILE_FLAG_NO_BUFFERING whenever you can instead.