@ethomson 2017-02-15
> The data git manages is too critical for corruption and/or failures.

I agree with @pureconfig here.  This is why we have a pluggable odb backend.  Our default is to mimic git; if they change the way they write object files, we will change to match.  But if you want some guarantees that git doesn't offer, then I would recommend putting your objects in a proper database.

> In particular, the way they batch changes, and test that fsync works seem like good ideas suitable for the odb backend.

I haven't read this paper, but my knee-jerk reaction is that any time we differ from git, we are incorrect.  Generally speaking, we aim for bug-for-bug compatibility with git.

> NOTE: do_fsync will always be default off, right?

Yes.  If git changes this default, so will we, but I can't imagine them doing so.