@ethomson 2017-02-15
>  Don't use FlushFileBuffers, it can be extremely slow (by Microsoft's own admission in the CreateFile docs). Use FILE_FLAG_WRITE_THROUGH | FILE_FLAG_NO_BUFFERING whenever you can instead.

/cc @whoisj

👋 I had it in my head that your filesystem team explicitly recommended that we use `FlushFileBuffers` here.  Do you happen to remember the details here?

If `FILE_FLAG_WRITE_THROUGH | FILE_FLAG_NO_BUFFERING` is safe enough, then I can use that instead.  Or we may be able to have a multistate setting that is unsafe and fast, reasonably safe and slow, and safest and punishing.