@pureconfig 2017-02-15
@ioquatix 

> Even without fsync, it's good enough actually, but it's a nice feature to have to improve durability

But you also said

>  correct is more important than performance. 

So "nice feature" is an understatement. As mentioned by someone else, it sounds like you want guarantees implying the use of a proper database.