@ioquatix 2017-02-15
@ethomson Do you guys use `libgit2` in Spokes (DGit)?