@ethomson 2017-02-17
@ioquatix We use libgit2 and git both on our fileservers, which are all in Spokes.  I don't want to spend a lot of time discussing GitHub's products in the libgit2 issue tracker, but the point of Spokes itself is to deal with problems like this.  You can learn more at https://www.youtube.com/watch?v=f7ecUqHxD7o and https://www.youtube.com/watch?v=DY0yNRNkYb0