@whoisj 2017-02-17
@ethomson the file system experts advise to *not* use the `FILE_FLAG_WRITE_THROUGH | FILE_FLAG_NO_BUFFERING` as they will likely cause exactly the behavior you're trying to avoid by emulating `fsync()` for Windows. They asked me to share this blog post: https://blogs.msdn.microsoft.com/oldnewthing/20140306-00/?p=1583/ .

Per the experts, the extra time spent not buffering and forcing the write to block on IO availability, almost guarantees your data won't be get to persistent media if there's a catastrophe. Instead, the recommend that the application flush data at appropriate times.

To quote them exactly:

> What is the point in having the data durable on the disk without the metadata durable?  So the bits are there, but after the power failure the file isn’t readable.
> 
> `FILE_FLAG_NO_BUFFERING` => Make it really slow to read the data later because it is not buffered.
> `FILE_FLAG_WRITE_THROUGH` => Make it really slow to write the data now because the I/O can’t be scheduled efficiently.
> 
> If the customer wants a flush they should use the API for flush.