@ethomson 2017-02-17
Thanks @whoisj for the definitive answer.

I've updated this to `fsync` parent directories in most cases (when a directory entry needs updating), except on Windows (where such semantics aren't permitted and don't make sense).

I've also changed the name of the option to `GIT_OPT_ENABLE_SYNCHRONOUS_OBJECT_CREATION` which makes more sense, I think.