@ethomson 2017-02-28
I added support for `core.fsyncObjectFiles`; this involved a little refactoring around how `odb`s get created.  I cribbed the pattern from `git_index_set_caps`, which I sort of loathe, but it's effective enough and an existing pattern.

This seems to be failing on Linux, though, after that.  I'll take a look.